fix(sdk-v4): renamed client.class.ts to client.ts
